Strip out non-prefixAll functionality.
Summary:
I removed the non-prefixAll functions and the corresponding tests.
Then, from each of the transformations I got rid of any browser checking that
it did. I also fixed a really annoying bug in `transition` which I encountered
in our aphrodite tests, and wrote a test for it.

The size of dist/prefixer.min.js shrunk from 9449 bytes to 6091 bytes in the
process.

Test Plan: - `npm run test`

Reviewers: jlfwong

Reviewed By: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D25206